SQLAlchemy for PMPermit and Snips

+++ stdplugins/snip.py
@@ -1,8 +1,14 @@
 # This Source Code Form is subject to the terms of the Mozilla Public
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
+"""Snips
+Available Commands:
+.snips
+.snipl
+.snipd"""
 from telethon import events, utils
 from telethon.tl import types
+from sql_helpers.snips_sql import get_snips, add_snip, remove_snip, get_all_snips
 
 
 TYPE_TEXT = 0
@@ -10,21 +16,23 @@ TYPE_PHOTO = 1
 TYPE_DOCUMENT = 2
 
 
-# {name: {'text': text, 'id': id, 'hash': access_hash, 'type': type}}
-snips = storage.snips or {}
-
-
-@borg.on(events.NewMessage(pattern=r'\.#(\S+)', outgoing=True))
+@borg.on(events.NewMessage(pattern=r'\#(\S+)', outgoing=True))
 async def on_snip(event):
     name = event.pattern_match.group(1)
-    if name not in snips:
-        await event.edit("This not does not exist")
-    else:
-        snip = snips[name]
-        if snip['type'] == TYPE_PHOTO:
-            media = types.InputPhoto(snip['id'], snip['hash'], snip['fr'])
-        elif snip['type'] == TYPE_DOCUMENT:
-            media = types.InputDocument(snip['id'], snip['hash'], snip['fr'])
+    snip = get_snips(name)
+    if snip:
+        if snip.snip_type == TYPE_PHOTO:
+            media = types.InputPhoto(
+                int(snip.media_id),
+                int(snip.media_access_hash),
+                snip.media_file_reference
+            )
+        elif snip.snip_type == TYPE_DOCUMENT:
+            media = types.InputDocument(
+                int(snip.media_id),
+                int(snip.media_access_hash),
+                snip.media_file_reference
+            )
         else:
             media = None
         message_id = event.message.id
@@ -32,7 +40,7 @@ async def on_snip(event):
             message_id = event.reply_to_msg_id
         await borg.send_message(
             event.chat_id,
-            snip['text'],
+            snip.reply,
             reply_to=message_id,
             file=media
         )
@@ -44,7 +52,6 @@ async def on_snip_save(event):
     name = event.pattern_match.group(1)
     msg = await event.get_reply_message()
     if msg:
-        snips.pop(name, None)
         snip = {'type': TYPE_TEXT, 'text': msg.message or ''}
         if msg.media:
             media = None
@@ -58,28 +65,37 @@ async def on_snip_save(event):
                 snip['id'] = media.id
                 snip['hash'] = media.access_hash
                 snip['fr'] = media.file_reference
-        snips[name] = snip
-        storage.snips = snips
-    await event.edit("Note {name} saved successfully. Get it with #{name}".format(name=name))
+        add_snip(name, snip['text'], snip['type'], snip.get('id'), snip.get('hash'), snip.get('fr'))
+    await event.edit("snip {name} saved successfully. Get it with #{name}".format(name=name))
 
 
 @borg.on(events.NewMessage(pattern=r'\.snipl', outgoing=True))
 async def on_snip_list(event):
-    await event.edit('available snips: ' + ', '.join(snips.keys()))
+    all_snips = get_all_snips()
+    OUT_STR = "Available Snips:\n"
+    if len(all_snips) > 0:
+        for a_snip in all_snips:
+            OUT_STR += f"👉 #{a_snip.snip} \n"
+    else:
+        OUT_STR = "No Snips. Start Saving using `.snips`"
+    if len(OUT_STR) > Config.MAX_MESSAGE_SIZE_LIMIT:
+        with io.BytesIO(str.encode(OUT_STR)) as out_file:
+            out_file.name = "snips.text"
+            await borg.send_file(
+                event.chat_id,
+                out_file,
+                force_document=True,
+                allow_cache=False,
+                caption="Available Snips",
+                reply_to=event
+            )
+            await event.delete()
+    else:
+        await event.edit(OUT_STR)
 
 
 @borg.on(events.NewMessage(pattern=r'\.snipd (\S+)', outgoing=True))
 async def on_snip_delete(event):
     name = event.pattern_match.group(1)
-    snips.pop(name, None)
-    storage.snips = snips
-    await event.edit("Note {} deleted successfully".format(name))
-
-
-@borg.on(events.NewMessage(pattern=r'\.snipr (\S+)\s+(\S+)', outgoing=True))
-async def on_snip_rename(event):
-    snip = snips.pop(event.pattern_match.group(1), None)
-    if snip:
-        snips[event.pattern_match.group(2)] = snip
-        storage.snips = snips
-    await event.delete()
+    remove_snip(name)
+    await event.edit("Note #{} deleted successfully".format(name))
